+++ desktop/frontend/bindings/github.com/wailsapp/wails/v3/internal/eventcreate.ts
@@ -6,17 +6,9 @@
 // @ts-ignore: Unused imports
 import { Create as $Create } from "@wailsio/runtime";
 
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore: Unused imports
-import * as main$0 from "../../../../BenedictKing/ccx/desktop/models.js";
-
 function configure() {
     Object.freeze(Object.assign($Create.Events, {
-        "update:available": $$createType0,
     }));
 }
-
-// Private type creation functions
-const $$createType0 = main$0.UpdateInfo.createFrom;
 
 configure();